# back-end/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://10.66.4.47:8989",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/staff/users/{id}/services": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get services assigned to a user",
                "operationId": "aea0fe32a719b153598caaf8d7b66133",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of user services",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UserResource"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Sync services assigned to a user",
                "description": "Replaces all existing service assignments with the provided list",
                "operationId": "f3da816f39a1f2c7b60e348d47a87a21",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SyncUserServicesRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Services synced successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UserResource"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "SyncUserServicesRequest": {
                "required": [
                    "service_ids"
                ],
                "properties": {
                    "service_ids": {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "example": 1
                        },
                        "example": [
                            1,
                            2,
                            3
                        ]
                    }
                },
                "type": "object"
            },
...
            "UserResource": {
                "allOf": [
                    {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "name": {
                                "type": "string",
                                "example": "John Doe"
                            },
                            "email": {
                                "type": "string",
                                "format": "email",
                                "example": "john@example.com"
                            },
                            "phone": {
                                "type": "string",
                                "example": "+359888123456"
                            },
                            "gender": {
                                "type": "string",
                                "enum": [
                                    "male",
                                    "female",
                                    "other"
                                ],
                                "example": "male"
                            },
                            "birth_date": {
                                "type": "string",
                                "format": "date",
                                "example": "1990-01-15"
                            },
                            "role": {
                                "type": "string",
                                "enum": [
                                    "admin",
                                    "worker",
                                    "client"
                                ],
                                "example": "worker"
                            },
                            "is_active": {
                                "type": "boolean",
                                "example": true
                            },
                            "is_approved": {
                                "type": "boolean",
                                "example": true
                            },
                            "email_verified": {
                                "type": "boolean",
                                "example": true
                            },
                            "job_position_id": {
                                "type": "integer",
                                "example": 1,
                                "nullable": true
                            }
                        },
                        "type": "object"
                    },
                    {
                        "$ref": "#/components/schemas/Timestamps"
                    }
                ]
            },
            "Timestamps": {
                "properties": {
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}